+++ Demo/Demo/Models/ClientIDResponse.swift
@@ -1,0 +1,8 @@
+struct ClientIDResponse: Codable {
+
+    enum CodingKeys: String, CodingKey {
+        case clientID = "value"
+    }
+
+    let clientID: String
+}