+++ Sources/CardPayments/APIRequests/ConfirmPaymentSourceRequest.swift
@@ -8,15 +8,14 @@ struct ConfirmPaymentSourceRequest: APIRequest {
     
     private let orderID: String
     private let pathFormat: String = "/v2/checkout/orders/%@/confirm-payment-source"
-    private let accessToken: String
+    private let base64EncodedCredentials: String
     var jsonEncoder: JSONEncoder
     
     /// Creates a request to attach a payment source to a specific order.
     /// In order to use this initializer, the `paymentSource` parameter has to
     /// contain the entire dictionary as it exists underneath the `payment_source` key.
-    
     init(
-        accessToken: String,
+        clientID: String,
         cardRequest: CardRequest,
         encoder: JSONEncoder = JSONEncoder() // exposed for test injection
     ) throws {
@@ -34,8 +33,7 @@ struct ConfirmPaymentSourceRequest: APIRequest {
         confirmPaymentSource.paymentSource = PaymentSource(card: card)
         
         self.orderID = cardRequest.orderID
-        self.accessToken = accessToken
-        
+        self.base64EncodedCredentials = Data(clientID.appending(":").utf8).base64EncodedString()
         path = String(format: pathFormat, orderID)
         
         jsonEncoder.keyEncodingStrategy = .convertToSnakeCase
@@ -60,7 +58,7 @@ struct ConfirmPaymentSourceRequest: APIRequest {
     var headers: [HTTPHeader: String] {
         [
             .contentType: "application/json", .acceptLanguage: "en_US",
-            .authorization: "Bearer \(accessToken)"
+            .authorization: "Basic \(base64EncodedCredentials)"
         ]
     }